chore(destination-mongodb): remove strict-encrypt variant, use env-aware TLS enforcement
- Add spec() override to MongodbDestination that removes TLS option for standalone instances when DEPLOYMENT_MODE=CLOUD
- Add TLS enforcement check in check() method for cloud deployment mode
- Remove registryOverrides.cloud.dockerRepository pointing to strict-encrypt
- Delete destination-mongodb-strict-encrypt connector directory

This migration allows the base connector to handle both OSS and Cloud deployments,
enforcing TLS in Cloud mode without needing a separate strict-encrypt connector.
