--- airbyte-integrations/connectors/destination-mongodb-strict-encrypt/README.md
@@ -1,1 +1,0 @@
-../_shared/README-java-destinations.md
