--- airbyte-integrations/connectors/destination-mongodb-strict-encrypt/build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id 'airbyte-java-connector'
-    id "io.airbyte.gradle.docker"
-    id 'airbyte-connector-docker-convention'
-}
-
-airbyteJavaConnector {
-    cdkVersionRequired = '0.19.0'
-    features = ['db-destinations', 'datastore-mongo']
-    useLocalCdk = false
-}
-
-application {
-    mainClass = 'io.airbyte.integrations.destination.mongodb.MongodbDestinationStrictEncrypt'
-    applicationDefaultJvmArgs = ['-XX:+ExitOnOutOfMemoryError', '-XX:MaxRAMPercentage=75.0']
-}
-
-dependencies {
-    implementation project(':airbyte-integrations:connectors:destination-mongodb')
-}
