--- airbyte-integrations/connectors/destination-mongodb-strict-encrypt/poe_tasks.toml
@@ -1,3 +1,0 @@
-include = [
-    "${POE_GIT_DIR}/poe-tasks/gradle-connector-tasks.toml",
-]
